chore: extract createDeferredPromise from fetch/utils.js
