+++ lib/web/fetch/body.js
@@ -4,7 +4,6 @@ const util = require('../../core/util')
 const {
   ReadableStreamFrom,
   readableStreamClose,
-  createDeferredPromise,
   fullyReadBody,
   extractMimeType,
   utf8DecodeBytes
@@ -17,6 +16,8 @@ const { isErrored, isDisturbed } = require('node:stream')
 const { isArrayBuffer } = require('node:util/types')
 const { serializeAMimeType } = require('./data-url')
 const { multipartFormDataParser } = require('./formdata-parser')
+const { createDeferredPromise } = require('../../util/promise')
+
 let random
 
 try {
