+++ lib/web/fetch/util.js
@@ -924,17 +924,6 @@ function sameOrigin (A, B) {
   return false
 }
 
-function createDeferredPromise () {
-  let res
-  let rej
-  const promise = new Promise((resolve, reject) => {
-    res = resolve
-    rej = reject
-  })
-
-  return { promise, resolve: res, reject: rej }
-}
-
 function isAborted (fetchParams) {
   return fetchParams.controller.state === 'aborted'
 }
@@ -1735,7 +1724,6 @@ module.exports = {
   isAborted,
   isCancelled,
   isValidEncodedURL,
-  createDeferredPromise,
   ReadableStreamFrom,
   tryUpgradeRequestToAPotentiallyTrustworthyURL,
   clampAndCoarsenConnectionTimingInfo,
